requirements.txt
streamlit
tensorflow==2.16.1
streamlit
torch
torchvision
Pillow